refactor(switch): move component and story file from tailwind folder
